+++ src/app/views/loan-detail/transaction-history/transaction-history.component.ts
@@ -49,7 +49,7 @@ export class TransactionHistoryComponent implements OnInit {
   timelinesProperties: object = {
     'loan_request': {
       'title': 'Requested',
-      'messege': 'Requested',
+      'message': 'Requested',
       'status': 'active',
       'materialClass': 'material-icons',
       'icon': 'code',
@@ -59,7 +59,7 @@ export class TransactionHistoryComponent implements OnInit {
     },
     'approved_loan': {
       'title': 'Loan Approved',
-      'messege': 'Loan Approved',
+      'message': 'Loan Approved',
       'status': 'active',
       'materialClass': 'material-icons',
       'icon': 'done',
@@ -69,7 +69,7 @@ export class TransactionHistoryComponent implements OnInit {
     },
     'lent': {
       'title': 'Lent',
-      'messege': 'Lent',
+      'message': 'Lent',
       'status': 'active',
       'materialClass': 'material-icons',
       'icon': 'trending_up',
@@ -79,7 +79,7 @@ export class TransactionHistoryComponent implements OnInit {
     },
     'partial_payment': {
       'title': 'Pay',
-      'messege': 'Pay',
+      'message': 'Pay',
       'status': 'active',
       'awesomeClass': 'fas fa-coins',
       'color': 'green',
@@ -88,7 +88,7 @@ export class TransactionHistoryComponent implements OnInit {
     },
     'total_payment': {
       'title': 'Completed',
-      'messege': 'Completed',
+      'message': 'Completed',
       'status': 'active',
       'awesomeClass': 'fas fa-check',
       'color': 'gray7',
@@ -97,7 +97,7 @@ export class TransactionHistoryComponent implements OnInit {
     },
     'loan_in_debt': {
       'title': 'In Debt',
-      'messege': 'In Debt',
+      'message': 'In Debt',
       'status': 'active',
       'materialClass': 'material-icons',
       'icon': 'error_outline',
@@ -107,7 +107,7 @@ export class TransactionHistoryComponent implements OnInit {
     },
     'withdraw': {
       'title': 'Withdraw',
-      'messege': 'Withdraw',
+      'message': 'Withdraw',
       'status': 'active',
       'materialClass': 'material-icons',
       'icon': 'call_made',
@@ -117,7 +117,7 @@ export class TransactionHistoryComponent implements OnInit {
     },
     'transfer': {
       'title': 'Transfer',
-      'messege': 'Transfer',
+      'message': 'Transfer',
       'status': 'active',
       'materialClass': 'material-icons',
       'icon': 'swap_horiz',
@@ -128,7 +128,7 @@ export class TransactionHistoryComponent implements OnInit {
     },
     'loan_expired': {
       'title': 'Expired',
-      'messege': 'Expired',
+      'message': 'Expired',
       'status': 'active',
       'materialClass': 'material-icons',
       'icon': 'snooze',
@@ -139,7 +139,7 @@ export class TransactionHistoryComponent implements OnInit {
     },
     'destroyed_loan': {
       'title': 'Destroyed',
-      'messege': 'Destroyed',
+      'message': 'Destroyed',
       'status': 'active',
       'materialClass': 'material-icons',
       'icon': 'delete',