+++ src/app/views/requested-loan/requested-loan.module.ts
@@ -1,5 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 // App Modules
 import { NgxSpinnerModule, NgxSpinnerService } from 'ngx-spinner';
 import { SharedModule } from '../../shared/shared.module';
@@ -7,6 +8,7 @@ import { MaterialModule } from './../../material/material.module';
 // App Services
 import { ContractsService } from './../../services/contracts.service';
 import { ApiService } from './../../services/api.service';
+import { FilterLoansService } from '../../services/filter-loans.service';
 // App Component
 import { RequestedLoanComponent } from './requested-loan.component';
 
@@ -15,7 +17,9 @@ import { RequestedLoanComponent } from './requested-loan.component';
     CommonModule,
     NgxSpinnerModule,
     MaterialModule,
-    SharedModule
+    SharedModule,
+    ReactiveFormsModule,
+    FormsModule
   ],
   declarations: [
     RequestedLoanComponent
@@ -23,7 +27,8 @@ import { RequestedLoanComponent } from './requested-loan.component';
   providers: [
     NgxSpinnerService,
     ContractsService,
-    ApiService
+    ApiService,
+    FilterLoansService
   ],
   exports: [
     RequestedLoanComponent