+++ CHANGELOG.md
@@ -1,10 +1,26 @@
 # Changelog
 
-## **Unreleased**
+## **0.0.12** Avocados - 2019/01/09
+
+### Features:
+- Notification Component
+
+### Misc:
+- Use Roboto font for number display
+- Increase extra ETH send from 3% to 5%
+- Increase min rebuy to 40 RCN
+
+### Fix:
+- Fix bug when view card-loan type visibility
+- Fix bug when approve contract checkbox visibility
+
+## **0.0.11** Avocados - 2018/12/25
 
 ### Features:
 - Added pay loan option
 - Lend using ETH if user has no RCN balance
+- Filter loan requests by currency, amount, interest rates and duration
+- Added Discord to Footer Component
 
 ### Misc:
 - Setup unit tests
@@ -14,6 +30,7 @@
 - Fix bug when loading metamask privacy on account detail
 - Fix try to lend an invalid mortgage request
 - Fix bug when the request is expired, the estimated return of a loan shows an incorrect amount
+- Fix margin "This address is not a lender" message
 
 ## **0.0.10-D** Avocados - 2018/11/14